backend_fastapi/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
